move test_body into lib
